- CalculateBidPerCapita now correctly calculating bids based on perception - CalculateBidPerCapita now clearer that it is based on some kind of perceived 'cost' of the good -Perceived cost can now be set -Perceived cost defaults to recipe cost and 'perceived cost' of good when asks don't exist in the market. - multipliers are now decimals

+++ Assets/EntityStrategies/BasicGrowthStrategy.cs
@@ -4,8 +4,8 @@ using System.Linq;
 public class BasicGrowthStrategy : iStrategy
 {
     //Manifestation of preferences
-    float _aggressionLevel = 0.2f;
-    public float GetAggressionLevel() => _aggressionLevel;
+    decimal _aggressionLevel = 0.2m;
+    public decimal GetAggressionLevel() => _aggressionLevel;
 
     public void GenerateGoals(iCompany company)
     {
@@ -60,7 +60,7 @@ public class BasicGrowthStrategy : iStrategy
         throw new System.NotImplementedException();
     }
 
-    public LemonadeStandResultObject SetAggressionLevel(float aggressionLevel)
+    public LemonadeStandResultObject SetAggressionLevel(decimal aggressionLevel)
     {
         _aggressionLevel = aggressionLevel;
         return LemonadeStandResultObject.Success();
